+++ lib/providers/consumer.dart
@@ -2,52 +2,60 @@ import 'package:flutter/material.dart';
 import 'package:module_provider/module_provider.dart';
 import 'package:module_provider/patterns/inherited_module.dart';
 
-class Consumer<T extends ChangeNotifier> extends StatefulWidget {
-  final T value;
-  final Widget Function(BuildContext context, Module module, T value) builder;
+abstract class Consumer<T extends ChangeNotifier, V> extends StatefulWidget {
+  final T provider;
+  final Widget Function(BuildContext context, V value) builder;
   
-  Consumer({Key key, @required this.value, @required this.builder}) : super(key: key);
+  Consumer({Key key, @required this.provider, @required this.builder}) : super(key: key);
 
-  T getValue(BuildContext context, ConsumerState consumer) {
-    return value;
+  T getProvider(BuildContext context, ConsumerState consumer) {
+    return provider;
   }
+  V getValue(BuildContext context, ConsumerState consumer);
 
   @override
-  ConsumerState createState() => ConsumerState<T>();
+  ConsumerState createState() => ConsumerState<T, V>();
 }
 
-class ConsumerState<T extends ChangeNotifier> extends State<Consumer<T>> {
-  Module module;
-  T value;
+class ConsumerState<T extends ChangeNotifier, V> extends State<Consumer<T, V>> {
+  Module _module;
+  Module get module => _module;
+
+  T _provider;
+  T get provider => _provider;
+
+  V _value;
 
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    
-    module = (context.inheritFromWidgetOfExactType(InheritedModule) as InheritedModule).module;
 
-    if (value == null) {
-      _initValue(context);
+    if (_provider == null) {
+      _initConsumer();
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    return widget.builder(context, module, value);
+    return widget.builder(context, _value);
   }
 
   @override
   void dispose() {
-    value.removeListener(_listener);
+    widget.provider.removeListener(_listener);
     super.dispose();
   }
 
-  _initValue(BuildContext context) {
-    value = widget.getValue(context, this);
-    value.addListener(_listener);
+  _initConsumer() {
+    _module = (context.inheritFromWidgetOfExactType(InheritedModule) as InheritedModule).module;
+    _provider = widget.getProvider(context, this);
+    _value = widget.getValue(context, this);
+    _provider.addListener(_listener);
   }
 
   _listener() {
-    setState(() {});
+    setState(() {
+      _value = widget.getValue(context, this);
+    });
   }
 }