+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+* Removed Module parameter from the ValueConsumer and ServiceConsumer.
+* Defined to consumers return direct the value in build function.
+* Added ValuesConsumer to consume ValuesProvider.
+* Updated lisense to "BSD 3-Clause License".
+
 ## 1.0.0
 
 * Initial version.